requirements.txt
coverage==4.4.1
flake8==3.3.0
isort==4.2.5
kivyunittest==0.1.8
eth-accounts==20191227
eth-hash==0.1.1
eth-utils==1.4.1
Kivy==1.11.1
kivy-garden.kivymd==2019.910
kivy-garden.qrcode==2019.914
kivy-garden.zbarcam==2019.1020
layoutmargin==20190911
Pillow==5.2.0
qrcode==5.3
raven==6.1.0
web3==5.2.0
